backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
sqlmodel==0.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==1.10.13 